Refactor code

+++ src/shader_program.hpp
@@ -23,7 +23,7 @@ public:
 	void setUniform(const std::string& name, bool value) const;
 	void setUniform(const std::string& name, int value) const;
 	void setUniform(const std::string& name, float value) const;
-	void setUniform(const std::string& name, const glm::vec2& value) const;
+	void setUniform(const std::string& name, const glm::ivec2& value) const;
 	void setUniform(const std::string& name, const glm::vec3& value) const;
 	void setUniform(const std::string& name, const glm::mat4& value) const;
